Fs 170 (#104)
* made file

* added IMU logs

* Added collection_timer setting and conditions to turn IMU on

* Added python imu plotting script

* added script file

* mid-debug

* made file

* added IMU logs

* Added collection_timer setting and conditions to turn IMU on

* Added python imu plotting script

* added script file

* mid-debug

* mid imu hardware debugging

* fixed initialization process for IMU, updated imu mode switching

* commit before merge

* shifted imu print statements to account for unset SensorReading Bounds

* removed collection_time from sfr (no longer needed)

* removed collection_time from other points in code (no longer needed)

* fixed formatting

* commit before switching branches

* Data from the IMU data collection test in the putty.log file

* Create universal sensor initialization type

* Debugging- IMU Monitor

* Reattempt imu initialization. Modified acceptable battery

* Added comment in python script

* Update git ignore and remove putty log

* Minor fixes in SFR and imu plotting script

---------

+++ src/Control Tasks/EEPROMControlTask.cpp
@@ -52,7 +52,7 @@ void EEPROMControlTask::save_sfr_data()
 
     if (time_since_last_write > sfr::eeprom::sfr_write_step_time) {
         // The last EEPROM write exceeds the interval between writes, so update the EEPROM value
-        int sfr_address = sfr::eeprom::sfr_address;
+        int sfr_address = (int)sfr::eeprom::sfr_address;
 
         /*
         > Each field is stored in this format: [boolean restore][T value].
@@ -81,7 +81,7 @@ void EEPROMControlTask::save_sfr_data()
 
         sfr::eeprom::sfr_address_age++;
         if (sfr::eeprom::sfr_address_age > 99000) { // Programmed write limit is less than the actual endurance of 100000 to create a safety buffer
-            sfr::eeprom::sfr_address += constants::eeprom::full_offset;
+            sfr::eeprom::sfr_address += (uint16_t)constants::eeprom::full_offset;
             EEPROM.put(5, sfr::eeprom::sfr_address);
         }
     }
